server/django_projects/requirements.txt
Django>=6.0.2
DjangoRestFramework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt
